Add trace_tensor op and refactor tensor tracing (#844)
We don't have an op that dispatches to the underlying
iree.turbine.ops.iree.trace_tensor.

With this change we can trace into safetensors files when executing both
in eager and with IREE.
Routes all tracing through the new tracing op. The user can set a
desired sink.

Added some functionality to prefix trace keys based on the module
structure.

Something that I could not preserve is the tensor trace counter. It is a
global variable and when exporting we pass multiple times through the
traced function, which makes the counter inconsistent. The only way to
distinguish traced tensors is through the trace keys.

The distinction between tensor tracing and goldens tracing has been
remove at the point of issuing a trace. If the user needs different
treatment they can customize the trace sink.