Make the T5 model tests use cosine similarity (#895)
There were several xfail tests with bad metric. Cosine similarity is a
better metric for language embeddings.

The comparison between bf16 and f32 exhibits a small fraction of
outliers that have a higher per-token numerical error than the majority
of tokens. To account for that the testing metric is expanded to test
for inlier and outlier absolute tolerance.