Decouple code for server startup, lifecycle hooks, FastAPI app, and routes (#962)
From an application standpoint, the code that we have to startup our
server, add routes, add lifecycle hooks, etc. is very tightly-coupled
and doesn't provide room for scale. It also makes it a bit difficult to
debug and write unit tests for the individual components.

This separates the logic for starting the server, creating the FastAPI
application, registering lifecycle hooks, and adding new routes.

It should make it pretty quick to expand upon each component
individually, and will allow us to add more features to the server
cleanly.

As a manual validation step (along with our CI), I ran my script that
sends 30 requests concurrently, results for that
[here](https://gist.github.com/stbaione/23673f7a8945b9b18eb7ffcfcebed8bc)