Add `argpartition` functionality to `shortfin.array` API (#1063)
This implements `argpartition` on the `shortfin.array` API, using
`xtensor`.

Argpartition is a sorting algorithm that returns indices, where all
indices to the left of k are guaranteed to be the k-smallest elements
along an axis, and where all values to the right of k are larger.

One can use a positive k-value, for the first k elements along an axis
to be the smallest k elements,
Or a negative k-value, for the last k elements along an axis to be the
largest k elements.

Note that those top-k indices are not guaranteed to be in sorted order.