BasePagedAttentionCache ref counting (#1089)
## Description

We need to add `ref_counting` to the pages in `BasePagedAttentionCache`
to support beam_search, sampling, etc.

When we `replicate` an `LlmInferenceExecRequest`, we will want to `fork`
the kvcache pages. The way this works is if `req_1` has n pages, and we
are forking the cache pages to `req_2`, then we will `share` cache pages
`1, ..., n-1`, and `copy` the nth page to a new page. This is because
pages `1, ..., n-1` will only be read from throughout the rest of the
generation process, while the last `incomplete` page may be written too.

When we `free` pages from an `LlmInferenceExecRequest`, we will want to
ensure that no other reqs are sharing any pages. If any pages are
shared, they should not be freed until their reference count hits `0`.
The pages that are not being shared (ref_count is already 0), can be
freed immediately.

Special attention to the use of the `threading.Lock` is appreciated.