Doc fixes for llm server 3.3 release (#1129)
There are three documentation related fixes that were identified through
QA testing:

- Specify `use-attention-mask` when running with a sharded model
- There's a special `--use-attention-mask` flag required to be able to
run the sharded models. I had forgotten to add that flag to the sharded
docs. This is what caused the compilation issue that Dhiraj had seen.
    - Context from Rob:
- This is due to some affinity analysis and cloning work that still
needs to be worked out. Its the main reason compilation work on iree
presubmits. There are plans to get a proper fix in but it will be a
while before someone can put hands on the issue.
- Removed deprecated `IREE` flag
- The `--iree-hal-target-backends=rocm` flag has been deprecated for
`--iree-hal-target-device=hip`. Removed the old flag in favor of the new
one.
- Changed instructions in `nightlies` to use `shortfin[apps]`
installation
- Dhiraj was doing his testing this morning with the nightlies, and hit
a bug due to `dataclasses_json` not being installed. The issue was that
he was downloading the nightly `shortfin`, not `shortfin[apps]`.
- Changed `nightly shortfin` instructions to install `shortfin[apps]`

---------

Co-authored-by: Scott Todd <scott.todd0@gmail.com>