[sharktank] Fix and enable sharded toy llama IREE test (#1211)
Check cosine similarity for the logits and the cache state instead of
element-wise comparison.

Make the test respect IREE target test args, allowing it to run not just
on CPU.

Make the test CLI accept multiple IREE devices.

Refactor get_iree_devices to accept multiple devices and to allow for
repeating devices even when not using "local-task". It will repeat the
devices to get to the desired device count. This is useful when you want
a test to run on a machine with fewer devices than required.
Allow for specifying device_count even when specifying a list of
devices.