Remove `temperature` bottleneck (#1276)
Previously, we were taking temperature on the entire `result_logits`
from a decode invocation.

Our `mistral` logits are of shape `[1, 1, 128256]`. When I ran a local
benchmark on how long the `sfnp.divide` function was taking for this
size of an array, it was ~14ms.

So, `(14ms * 16 parallel_reqs * 64 decode_steps) / 1000 == 14 s`!

In local benchmarks, I was able to decrease the latency of 16 concurrent
requests from `~25s` to `~9s`, which also increased our throughput quite
a bit.

The idea is pretty simple:

For greedy selection, or if already in a `softmax`/`log_softmax` form
don't take temperature. Dividing by a scalar won't impact which token is
the highest scoring token or which ones are the `top_k` highest scoring
tokens. It doesn't make sense to apply temperature if already in
`softmax` or `log_softmax` form.

Otherwise only take temperature on the values that are getting converted
to `softmax`. This reduces the size of the array we are dividing
drastically.