[sharktank] Dense FFN expert block in MoE (#1359)
The current PreGatherFFNMOE would index into the expert weights with the
selected top experts. Resulting in a large tensor that IREE does not
have a way to not materialize it and this will blow up the memory.
```
        weights = weights[experts, :, :]
```
The shape of experts is `(batch_size * sequence_length,
num_top_experts)`. For large `batch_size * sequence_length` the
resulting matrix would become quite big. For Deepseek V3 for example we
would get `(batch_size * sequence_length, num_top_experts, 2048, 7168)`.

Per the suggestion of Rob Suderman and when looking how Hugging Face's
Llama 4 implements MoE, here is introduced a Dense FFN expert block. It
routes all tokens to all experts no matter the number of top experts. I
don't know how this could be faster than a sparse block as it does a
bunch of useless computations, but it reduces the number of high-level
ops as it does not loop through the experts. We could at least try it.
It is supposed to perform a bit better with expert-parallel sharding.