[sharktank] General implementation for trivially replicable ops (#1384)
The majority of operations have essentially the same replication
mechanism across device. We call the underlying unsharded op with the
corresponding shards and then construct a replicated tensor form the
resulting shards.

This change solves this in general way to not have to write the same
replication pattern for all ops.