Shortfin gpu topk (#1474)
## Description

Stacked on #1422 

This PR wires up the `topk` kernel from `sharktank` through the
`shortfin LLM server`.

Similarly to `argmax`, if the model is exported with `topk > 1`, it will
return the indices and subset of values that correspond to the `k`
highest scoring logits, in sorted order.

If an invalid combination of `SeverParams or GenReqInput` with
`ModelParams` is used, the server returns a `400` error, with this body:

```json
{
    "error": "Requested top-k larger than exported top-k",
    "code": "INVALID_REQUEST_ARGS",
    "exported_topk": 4,
    "requested_topk": 8
}
```

We won't see any performance improvements from using the gpu based
`topk` route, until the [codegen performance is
fixed](https://github.com/iree-org/iree/issues/20772)