[sharktank] In trace tensor module patching allow for exclusion by regex key search (#1617)
In some strange cases torch does not want to trace tensors during
export.
It complains about that we are trying to modify frozen tensors. We need
our signature to be
```
trace_tensor(str trace_key, Tensor(a!) tensor) -> ()
```
This means it modifies the tensor in-place and returns nothing. Then the
op
will not be removed.
Cloning the tensor before tracing does not help as torch export is smart
and
removes the whole expression.
Ideally we would like to say that the op is side-effecting but that
seems
unavailable in torch.

With this change you can exclude some trace key patterns to be excluded
from the trace.
Thus circumventing the torch weirdness.