[sharktank] allow for different dtypes and sort by file modification time in assert_close_safetensors (#1618)
Sometimes when comparing traces the reference may be of higher
precision. Here a flag is added to allow for this auto promotion.

Sort by file modification time to try and get an order closer to the
reference execution order. When we trace, tensors are written in order
as the module is executed and we want to preserve this order when doing
the comparison. It would be ideal if we could have a global trace tensor
counter. Then we would be able to do a lexicographical sort, but this
does not play nice during torch export.