Enable client-side `num_beams` and `beam_search` settings (#1558)
Stacked on #1546 

# Description

This enables the ability to switch the number of beams, and between
`default` and `beam_search` sampling through the client.

If desired, a user can still define a `server default` where they set a
default number of beams and whether or not to use beam_search. All
requests use this server default unless otherwise specified from the
client request.

The client can specify two fields:
- num_beams (int): The number of beams to use for generation
- use_beam_search (bool): Whether or not to use `beam_search`

This also fixes a bug in the rate limiting behavior, introduced by
#1275. When calculating the number of requests to add to the queue, it
wasn't properly taking into account that a single http request may
contain a batch input (multiple completion requests). It had to be
adjusted for this feature anyways.

Added smoke tests for all of the combinations.