Change dtypes for toy models (#1693)
Addresses a closely few issues:
1. Deepseek was using fp32 weights instead of fp16 weights.
2. Toy models were incorrectly using the same dtype for the norm weights
as for the other layers. E.g. llama8b using fp32 for ffn_norm and
attn_norm but fp16 for everything else, currently implementation was
using fp16 for all.

It also expands the deepseek tests to:
1. Use KL-divergence for comparing logit distributions rather than a
direct 1-to-1 comparisons of each logits.
2.  Paramaterize on dtype.

Finally, it splits out a helper method in `TorchGenerator` to seperately
preprocess tokens as the same operation is copy and pasted across
countless tests.

---------

Signed-off-by: Alex Vasile <48962821+Alex-Vasile@users.noreply.github.com>