Add unit tests for Llm Batcher classes (#1643)
# Description

1. Unit tests for Batcher classes
2. Very small change to make `LlmExecutorProcess` more testable by
moving the post-processing calls to a `_post_run` function.
- Also added some type hints and doc strings at parts that I found
confusing while writing tests.
4. Removes the `xfails` that we previously had on the `fork_pages` test.
I spent annoying amount of time trying to figure out how to let a test
wait on the device. It was obvious once I found it, and same trick fixes
`fork_pages`.

The batcher is likely the most critical component in the `Shortfin Llm
Server`, but we were lacking a set of unit tests for it.
We do have some `direct_to_batcher` tests that runs an integration test
in a simulated decode loop. However, it's better that we have a more
robust wall.

For the most part, I only test public functions that have non-trivial
logic. I did specifically chose to test two private functions:

- _transfer_buffers
- _post_run

Added a test for _transfer_buffers, because I've seen a couple PRs
attempt to refactor this specific spot incorrectly. So, it seems like
the easiest function to get wrong, but also the most enticing to change
haha.

Added a test for _post_run, because I really just created that function
to make the tests a little easier. Still includes important logic that
we want a guard around.