Re-xfail `test_fork_pages` (#1749)
For some reason, the `page_table` isn't being setup properly. When the
error is thrown that `New last page should be filled with 1s`, it's
actually because the page_table it's copying from isn't properly filling
the values before calling `fork_pages`.

Just adding an `await device0` didn't fix it, so need to look further. 

Either something wrong with test setup, or possible a bug in `.fill`